+++ src/scalars/components/url-field/url-field.stories.tsx
@@ -58,16 +58,6 @@ const meta: Meta<typeof UrlField> = {
     ...getValidationArgTypes(),
 
     ...PrebuiltArgTypes.viewMode,
-    diffMode: {
-      control: 'select',
-      description: 'The mode of the input field',
-      options: ['sentences'],
-      table: {
-        type: { summary: 'sentences' },
-        defaultValue: { summary: 'sentences' },
-        category: StorybookControlCategory.DIFF,
-      },
-    },
     ...PrebuiltArgTypes.baseValue,
   },
   args: {