+++ src/ui/components/data-entry/select/content.tsx
@@ -23,7 +23,6 @@ interface ContentProps {
   favoriteOptions?: SelectProps['options']
   toggleAll: () => void
   toggleOption: (value: string) => void
-  optionsClassName?: string
 }
 
 export const Content: React.FC<ContentProps> = ({
@@ -37,7 +36,6 @@ export const Content: React.FC<ContentProps> = ({
   toggleAll,
   toggleOption,
   favoriteOptions = [],
-  optionsClassName,
 }) => {
   const enabledOptions = options.filter((opt) => !opt.disabled)
   const hasAnyIcon = options.some((opt) => opt.icon)
@@ -49,7 +47,7 @@ export const Content: React.FC<ContentProps> = ({
   }, [commandListRef, cmdkSearch])
 
   return (
-    <>
+    <div className={cn('select__content')}>
       {searchable && (
         <CommandInput
           placeholder="Search..."
@@ -60,10 +58,10 @@ export const Content: React.FC<ContentProps> = ({
             }
           }}
           wrapperClassName="rounded-t"
-          className="text-gray-900 dark:text-gray-50"
+          className="text-gray-900 dark:text-gray-50 select__search"
         />
       )}
-      <CommandList ref={commandListRef} tabIndex={!searchable ? 0 : undefined}>
+      <CommandList ref={commandListRef} tabIndex={!searchable ? 0 : undefined} className="select__list">
         <CommandEmpty className="p-4 text-center text-[14px] font-normal leading-5 text-gray-700 dark:text-gray-400">
           No results found.
         </CommandEmpty>
@@ -121,7 +119,7 @@ export const Content: React.FC<ContentProps> = ({
             hasAnyIcon={hasAnyIcon}
             toggleOption={toggleOption}
             tabIndex={!searchable ? 0 : undefined}
-            optionsClassName={optionsClassName}
+            className="select__list-item-favorite"
           />
 
           {favoriteOptions.length > 0 && <div className="my-1 border-b border-gray-300 dark:border-gray-600" />}
@@ -134,10 +132,10 @@ export const Content: React.FC<ContentProps> = ({
             hasAnyIcon={hasAnyIcon}
             toggleOption={toggleOption}
             tabIndex={!searchable ? 0 : undefined}
-            optionsClassName={optionsClassName}
+            className="select__list-item"
           />
         </CommandGroup>
       </CommandList>
-    </>
+    </div>
   )
 }