+++ src/ui/components/object-set-table/subcomponents/header.tsx
@@ -1,4 +1,4 @@
-import { type FC, useCallback } from 'react'
+import { type FC, useCallback, useMemo } from 'react'
 import { cn } from '../../../../scalars/lib/utils.js'
 import type { CellContext, ColumnDef } from '../types.js'
 import { getColumnTitle } from '../utils.js'
@@ -24,22 +24,53 @@ const TableHeader: FC<TableHeaderProps> = ({ columns }) => {
 
   const isAllRowsSelected = selectedRowIndexes.length === config.data.length
 
+  const columnHeaders = useMemo(() => {
+    return columns.map((column, columnIndex) => {
+      const cellContext: CellContext<unknown> = {
+        rowIndex: -1, // the header row has no row
+        row: undefined,
+        column,
+        columnIndex,
+        tableConfig: config,
+      }
+
+      const style: React.CSSProperties = {
+        width: column.width ?? 'auto',
+        minWidth: column.minWidth ?? 'auto',
+        maxWidth: column.maxWidth ?? column.width ?? 'auto',
+      }
+
+      const onSort = () => {
+        if (!column.sortable) return
+
+        const tableState = api._getState()
+        if (tableState.sortState?.columnIndex === columnIndex) {
+          const nextDirection = tableState.sortState.direction === 'asc' ? 'desc' : null
+          api.sortRows(columnIndex, nextDirection)
+        } else {
+          api.sortRows(columnIndex, 'asc')
+        }
+      }
+
+      return (
+        <th
+          className={cn('group/header-cell', column.sortable ? 'cursor-pointer' : '')}
+          style={style}
+          key={column.field}
+          onClick={onSort}
+        >
+          {/* the `renderHeader` function should be intialized with a default value at this point */}
+          {column.renderHeader!(getColumnTitle(column), cellContext)}
+        </th>
+      )
+    })
+  }, [api, columns, config])
+
   return (
     <thead>
       <tr className={cn('border-gray-300', !selectedRowIndexes.includes(0) && 'border-b')}>
         <HeaderNumberTd isAllRowsSelected={isAllRowsSelected} handleSelectAllRows={handleSelectAllRows} />
-        {columns.map((column, columnIndex) => {
-          const cellContext: CellContext<unknown> = {
-            rowIndex: -1, // the header row has no row
-            row: undefined,
-            column,
-            columnIndex,
-            tableConfig: config,
-          }
-
-          // the `renderHeader` function should be intialized with a default value at this point
-          return <th key={column.field}>{column.renderHeader!(getColumnTitle(column), cellContext)}</th>
-        })}
+        {columnHeaders}
       </tr>
     </thead>
   )